[BUG #31300] check the accessibility, only admin or owner can access/delete file

+++ src/main/java/com/ensemble/entreprendre/service/impl/FileDbServiceImpl.java
@@ -5,9 +5,12 @@ import org.springframework.stereotype.Service;
 
 import com.ensemble.entreprendre.converter.GenericConverter;
 import com.ensemble.entreprendre.domain.FileDb;
+import com.ensemble.entreprendre.domain.User;
 import com.ensemble.entreprendre.dto.FileDbDto;
+import com.ensemble.entreprendre.exception.ApiException;
 import com.ensemble.entreprendre.exception.ApiNotFoundException;
 import com.ensemble.entreprendre.repository.IFileDbRepository;
+import com.ensemble.entreprendre.service.IConnectedUserService;
 import com.ensemble.entreprendre.service.IFileDbService;
 
 @Service
@@ -19,15 +22,22 @@ public class FileDbServiceImpl implements IFileDbService {
 	@Autowired
 	GenericConverter<FileDb, FileDbDto> fileDbConverter;
 
+	@Autowired
+	IConnectedUserService connectedUserService;
+
 	@Override
-	public FileDbDto getById(Long id) throws ApiNotFoundException {
+	public FileDbDto getById(Long id) throws ApiException {
 		FileDb file = getFileIfExists(id);
+		this.checkAccess(file.getUser());
+
 		return this.fileDbConverter.entityToDto(file, FileDbDto.class);
 	}
 
 	@Override
-	public void deleteFile(Long id) throws ApiNotFoundException {
+	public void deleteFile(Long id) throws ApiException {
 		FileDb file = getFileIfExists(id);
+		this.checkAccess(file.getUser());
+
 		this.fileDbRepository.delete(file);
 	}
 
@@ -35,6 +45,15 @@ public class FileDbServiceImpl implements IFileDbService {
 		FileDb file = this.fileDbRepository.findById(id)
 				.orElseThrow(() -> new ApiNotFoundException("Ce fichier n'existe pas !"));
 		return file;
+	}
+
+	/*
+	 * Throw AccessDeniedException if the user does not have the permissions
+	 */
+	private void checkAccess(User user) throws ApiException {
+		if (!connectedUserService.isAdmin()) {
+			this.connectedUserService.isOwner(user);
+		}
 	}
 
 }
